Return HttpStatusCode 405 in case method is not implemented
Also rename helper to controller and split test files into small files.
