+++ controller.js
@@ -1,6 +1,6 @@
 const connector = require('./connector');
 
-helper = {
+module.exports = {
   'has_authorization_key': (req) => {
     if (req.headers['authorization'] === undefined) {
       return false;
@@ -13,5 +13,3 @@ helper = {
     });
   }
 };
-
-module.exports = helper;
