+++ routes.js
@@ -1,10 +1,10 @@
 const routes = require('express').Router();
-const helper = require('./helper');
+const controller = require('./controller');
 
 routes.route('/')
   .get((req, res) => {
-    if (helper.has_authorization_key(req)) {
-      helper.is_authorized(req, function(err, authorized) {
+    if (controller.has_authorization_key(req)) {
+      controller.is_authorized(req, function(err, authorized) {
         if (authorized === false) {
           res.status(400).json({
             statusCode: 400,
@@ -24,6 +24,12 @@ routes.route('/')
         message: "You need to insert a token as 'authorization' on Header"
       });
     };
+  })
+  .all((req, res) => {
+    res.status(405).json({
+      statusCode: 405,
+      message: "The requested method is not allowed!"
+    });
   });
 
 routes.route('/*')
